refactor: Update AccountPopup and SignInPopup components
The code changes in this commit refactor the `AccountPopup.svelte` and `SignInPopup.svelte` components.

In `AccountPopup.svelte`, the `getSettings` function from the `$lib/utils/settings` module is imported, but commented out. The purpose of this change is not clear from the code alone.

In `SignInPopup.svelte`, the `addUserDoc` function has been modified to remove a TODO comment. The purpose of this change is not clear from the code alone.

In `UserSignInIcon.svelte`, the `toggleDropdown` and `togglePopup` functions have been modified to be asynchronous. The purpose of this change is not clear from the code alone.

This commit message follows the convention of starting with a type prefix (`refactor`) to indicate the nature of the change, followed by a concise and descriptive message.